move string_processor under the core directory
