+++ flexeval/__init__.py
@@ -10,7 +10,6 @@ from .core.few_shot_generator import *
 from .core.generation_dataset import *
 from .core.language_model import *
 from .core.metric import *
-from .core.metric.string_processor import *
 from .core.metric.tokenizer import *
 from .core.multiple_choice_dataset import *
 from .core.pairwise_comparison import *
@@ -18,5 +17,6 @@ from .core.prompt_template import *
 from .core.result_recorder import *
 from .core.reward_bench_dataset import *
 from .core.reward_model import *
+from .core.string_processor import *
 from .core.text_dataset import *
 from .utils import *
