+++ flexeval/core/metric/char_f1.py
@@ -4,8 +4,9 @@ import functools
 
 from fuzzywuzzy import fuzz
 
+from flexeval.core.string_processor import StringProcessor
+
 from .base import Metric, MetricResult
-from .string_processor import StringProcessor
 
 
 class CharF1(Metric):
