+++ flexeval/core/string_processor/string_strip.py
