+++ tests/core/metric/string_processor/test_nfkc.py
@@ -1,6 +1,6 @@
 import pytest
 
-from flexeval.core.metric.string_processor import NFKCNormalizer
+from flexeval.core.string_processor import NFKCNormalizer
 
 
 @pytest.mark.parametrize(
