+++ tests/core/metric/string_processor/test_string_strip.py
@@ -1,6 +1,6 @@
 import pytest
 
-from flexeval.core.metric.string_processor import StringStrip
+from flexeval.core.string_processor import StringStrip
 
 
 @pytest.mark.parametrize(
